ticket_agent/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-anthropic>=0.0.2
python-dotenv>=1.0.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
tokenizers>=0.21